Aplicacion de validaciones con Express-Validator

+++ src/views/cursos.js
@@ -1,9 +1,7 @@
 document.getElementById("modal-modificar").style.display = "none";
 
-// Se ejecuta cuando el documento está completamente cargado
 document.addEventListener("DOMContentLoaded", () => {
   // Cargar la lista de cursos
-
   obtenerCursos();
 
   // Cargar la lista de profesores en el formulario de agregar curso
@@ -12,23 +10,73 @@ document.addEventListener("DOMContentLoaded", () => {
   // Evento para manejar el envío del formulario de agregar curso
   const formCurso = document.getElementById("form-curso");
   formCurso.addEventListener("submit", async (event) => {
-    event.preventDefault(); // Evitar que se recargue la página
-    const nombre = document.getElementById("nombre").value; // Obtener el nombre del curso
-    const descripcion = document.getElementById("descripcion").value; // Obtener la descripción del curso
-    const profesor_id = document.getElementById("profesor").value; // Obtener el id del profesor seleccionado
+    event.preventDefault();
+    const nombre = document.getElementById("nombre").value;
+    const descripcion = document.getElementById("descripcion").value;
+    const profesor_id = document.getElementById("profesor").value;
 
-    // Crear un objeto con los datos del curso nuevo
-    const nuevoCurso = { nombre, descripcion, profesor_id };
-    // Llamar a la función agregarCurso, que se encarga de enviarlo al backend
-    await agregarCurso(nuevoCurso);
-    formCurso.reset(); // Reiniciar el formulario después de agregar
+    const validacionExitosa = validarCurso(nombre, descripcion, profesor_id);
+    if (!validacionExitosa) {
+      return;
+    } else {
+      const nuevoCurso = { nombre, descripcion, profesor_id };
+      await agregarCurso(nuevoCurso);
+      formCurso.reset();
+    }
   });
 });
 
-// Obtener la lista de cursos desde el servidor
+/**** VALIDACION ****/
+function validarCurso(nombre, descripcion, profesor_id) {
+  const nombreVal = /^[a-zA-Z0-9\s]{3,50}$/;
+  const descripcionVal = /^.{5,200}$/;
+  const profesor_idVal = /^\d+$/;
+
+  if (!nombreVal.test(nombre) && !descripcionVal.test(descripcion)) {
+    Swal.fire({
+      icon: "error",
+      title: "Campos incompletos",
+      text: "Por favor, complete todos los campos.",
+    });
+    return false;
+  }
+
+  // Validar que el nombre no esté vacío y tenga al menos 3 caracteres
+  if (!nombreVal.test(nombre)) {
+    Swal.fire({
+      icon: "warning",
+      title: "Falta el Nombre",
+      text: "El nombre debe tener al menos 3 caracteres.",
+    });
+    return false;
+  }
+
+  // Validar que la descripción no esté vacía y tenga al menos 10 caracteres
+  if (!descripcionVal.test(descripcion)) {
+    Swal.fire({
+      icon: "warning",
+      title: "Falta la Descripción",
+      text: "La descripción debe tener al menos 10 caracteres.",
+    });
+    return false;
+  }
+
+  // Validar que se haya seleccionado un profesor
+  if (!profesor_idVal.test(profesor_id)) {
+    Swal.fire({
+      icon: "error",
+      title: "Profesor no seleccionado",
+      text: "Debe seleccionar un profesor para el curso.",
+    });
+    return false;
+  }
+
+  return true; // Si todas las validaciones pasan
+}
+
+/**** OBTENER LISTA DE CURSOS ****/
 async function obtenerCursos() {
   try {
-    // Realizar una solicitud GET al servidor
     const response = await fetch("http://localhost:3000/cursos", {
       method: "GET",
       headers: {
@@ -48,8 +96,8 @@ async function obtenerCursos() {
             <td>${curso.descripcion}</td>
             <td>${curso.profesor.nombre} ${curso.profesor.apellido}</td>
             <td class= "tabla-accion">
-                <button onclick="eliminarCurso(${curso.id})">Eliminar</button>
-                <button onclick="mostrarFormularioModificar(${curso.id})">Modificar</button>
+              <button class="btn-modificar" onclick="mostrarFormularioModificar(${curso.id})"><i class="fas fa-pencil-alt"></i></button>
+              <button class="btn-eliminar" onclick="eliminarCurso(${curso.id})"><i class="fas fa-times"></i></button>
             </td>
           `;
       tbody.appendChild(row); // Agregar la fila a la tabla
@@ -59,7 +107,7 @@ async function obtenerCursos() {
   }
 }
 
-// Obtener la lista de profesores para el formulario de agregar curso
+/**** OBTENER LISTA DE PROFESORES ****/
 async function obtenerProfesoresParaFormulario() {
   try {
     const response = await fetch("http://localhost:3000/profesores");
@@ -78,7 +126,7 @@ async function obtenerProfesoresParaFormulario() {
   }
 }
 
-// Agregar un nuevo curso al backend
+/**** AGREGAR CURSO ****/
 async function agregarCurso(curso) {
   try {
     const response = await fetch("http://localhost:3000/cursos", {
@@ -90,38 +138,67 @@ async function agregarCurso(curso) {
     });
 
     if (response.ok) {
-      alert("Curso agregado correctamente.");
+      Swal.fire({
+        icon: "success",
+        title: "Curso agregado",
+        text: "El curso ha sido agregado correctamente.",
+      });
       obtenerCursos(); // Volver a cargar la lista de cursos
     } else {
-      console.error("Error al agregar curso");
+      Swal.fire({
+        icon: "error",
+        title: "Error al agregar curso",
+        text: "Hubo un problema al agregar el curso.",
+      });
     }
   } catch (error) {
     console.error("Error al agregar curso:", error);
   }
 }
 
-// Eliminar un curso
+/**** ELIMINAR CURSO ****/
 async function eliminarCurso(id) {
-  try {
-    const response = await fetch(`http://localhost:3000/cursos/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
+  Swal.fire({
+    title: "¿Está seguro?",
+    text: "No podrá revertir esta acción. ¿Desea eliminar el curso?",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#145c17",
+    cancelButtonColor: "#6b1515",
+    confirmButtonText: "Sí, eliminar",
+    cancelButtonText: "Cancelar",
+  }).then(async (result) => {
+    if (result.isConfirmed) {
+      try {
+        const response = await fetch(`http://localhost:3000/cursos/${id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
 
-    if (response.ok) {
-      alert("Curso eliminado correctamente.");
-      obtenerCursos(); // Volver a cargar la lista de cursos después de eliminar
-    } else {
-      console.error("Error al eliminar curso");
+        if (response.ok) {
+          Swal.fire({
+            icon: "success",
+            title: "Curso eliminado",
+            text: "El curso ha sido eliminado exitosamente.",
+          });
+          obtenerCursos(); // Volver a cargar la lista de cursos después de eliminar
+        } else {
+          Swal.fire({
+            icon: "error",
+            title: "Error al eliminar curso",
+            text: "No se pudo eliminar el curso.",
+          });
+        }
+      } catch (error) {
+        console.error("Error al eliminar curso:", error);
+      }
     }
-  } catch (error) {
-    console.error("Error al eliminar curso:", error);
-  }
+  });
 }
 
-// Mostrar el formulario para modificar un curso
+/**** MODIFICAR CURSO ****/
 async function mostrarFormularioModificar(id) {
   try {
     const response = await fetch(`http://localhost:3000/cursos/${id}`);
@@ -161,7 +238,7 @@ async function mostrarFormularioModificar(id) {
   }
 }
 
-// Evento para manejar el envío del formulario de modificación
+// Manejar la modificación de un curso
 document
   .getElementById("form-modificar")
   .addEventListener("submit", async function (event) {
@@ -173,11 +250,14 @@ document
     const descripcion = document.getElementById("modificar-descripcion").value;
     const profesor_id = document.getElementById("modificar-profesor").value;
 
-    // Crear un objeto con los nuevos valores del curso
-    const cursoModificado = { nombre, descripcion, profesor_id };
-
-    // Llamar a la función que envía los nuevos datos al servidor
-    await modificarCurso(id, cursoModificado);
+    // Validar los campos del formulario de modificación
+    const validacionExitosa = validarCurso(nombre, descripcion, profesor_id);
+    if (!validacionExitosa) {
+      return;
+    } else {
+      const cursoModificado = { nombre, descripcion, profesor_id };
+      await modificarCurso(id, cursoModificado);
+    }
   });
 
 // Enviar la solicitud de modificación del curso al servidor
@@ -192,18 +272,26 @@ async function modificarCurso(id, cursoModificado) {
     });
 
     if (response.ok) {
-      alert("Curso modificado correctamente.");
+      Swal.fire({
+        icon: "success",
+        title: "Curso modificado",
+        text: "El curso ha sido modificado exitosamente.",
+      });
       document.getElementById("modal-modificar").style.display = "none"; // Cerrar el modal
       obtenerCursos(); // Recargar la lista de cursos
     } else {
-      console.error("Error al modificar curso");
+      Swal.fire({
+        icon: "error",
+        title: "Error al modificar curso",
+        text: "No se pudo modificar el curso.",
+      });
     }
   } catch (error) {
     console.error("Error al modificar curso:", error);
   }
 }
 
-// Cerrar el modal cuando el usuario haga clic fuera del modal
+// Cerrar el pop up cuando el usuario hace clic fuera
 window.onclick = function (event) {
   const modal = document.getElementById("modal-modificar");
   if (event.target === modal) {
@@ -211,7 +299,7 @@ window.onclick = function (event) {
   }
 };
 
-// Cerrar el modal cuando el usuario haga clic en la 'X'
+// Cerrar el pop up cuando el usuario haga clic en la 'X'
 document.querySelector(".close").onclick = function () {
   document.getElementById("modal-modificar").style.display = "none";
 };
